- fixed trafficdbcompact errors with empty stats database and fixed inlineing of vars in queries

+++ modules/trafficdbcompact.php
@@ -48,13 +48,15 @@ if($delete)
 
 if($removedeleted)
 {
-    $nodes_from_stats = $LMS->DB->GetCol("SELECT DISTINCT nodeid FROM stats"); 
-    $nodes = $LMS->DB->GetCol("SELECT id FROM nodes");
-    foreach($nodes_from_stats as $node)
+    if($nodes_from_stats = $LMS->DB->GetCol("SELECT DISTINCT nodeid FROM stats")) 
     {
-	if(!in_array($node,$nodes))
-	    if($LMS->DB->Execute("DELETE FROM stats WHERE nodeid = ".$node))
-		echo "Usuniêto statystyki komputera o ID: ".$node."\n";
+	$nodes = $LMS->DB->GetCol("SELECT id FROM nodes");
+	foreach($nodes_from_stats as $node)
+	{
+	    if(!in_array($node,$nodes))
+		if($LMS->DB->Execute("DELETE FROM stats WHERE nodeid = ".$node))
+		    echo "Usuniêto statystyki komputera o ID: ".$node."\n";
+	}
     }
 }
 
@@ -67,29 +69,31 @@ if($level)
 	case '2' : $period = $time-30*24*60*60; $step = 24*60*60; break;//mies, dzieñ
 	case '3' : $period = $time-365*24*60*60; $step = 60*60; break; //po miesi±c, godz	
     }
-    $mintime = $LMS->DB->GetOne("SELECT MIN(dt) FROM stats");
-    //$nodes = $LMS->GetNodeList("id,asc");
-    $nodes = $LMS->DB->GetAll("SELECT id, name FROM nodes ORDER BY name");
-    foreach($nodes as $node)
+    
+    if($mintime = $LMS->DB->GetOne("SELECT MIN(dt) FROM stats"))
     {
-        echo "'".$node['name']."'\t: "; 
-	$deleted = 0;
-	$inserted = 0;
-	$LMS->DB->BeginTrans();
-	$maxtime = $period;
-	while($maxtime > $mintime)
+	$nodes = $LMS->DB->GetAll("SELECT id, name FROM nodes ORDER BY name");
+	foreach($nodes as $node)
 	{
-	    $data = $LMS->DB->GetRow("SELECT sum(upload) as upload, sum(download) as download FROM stats WHERE dt >= $maxtime - $step AND dt < $maxtime AND nodeid=$node['id'] GROUP BY nodeid");
-	    $deleted += $LMS->DB->Execute("DELETE FROM stats WHERE nodeid=$node['id'] AND dt >= $maxtime - $step AND dt < $maxtime"); 
-	    $download = ($data['download']?$data['download']:0);
-	    $upload = ($data['upload']?$data['upload']:0);
-	    if($download || $upload)
-		$inserted += $LMS->DB->Execute("INSERT INTO stats (nodeid, dt, upload, download) VALUES ($node['id'], $maxtime, $upload, $download )");
-	    $maxtime -= $step;
-	}
+    	    echo "'".$node['name']."'\t: "; 
+	    $deleted = 0;
+	    $inserted = 0;
+	    $LMS->DB->BeginTrans();
+	    $maxtime = $period;
+	    while($maxtime > $mintime)
+	    {
+		$data = $LMS->DB->GetRow("SELECT sum(upload) as upload, sum(download) as download FROM stats WHERE dt >= $maxtime - $step AND dt < $maxtime AND nodeid=? GROUP BY nodeid", array($node['id']));
+		$deleted += $LMS->DB->Execute("DELETE FROM stats WHERE nodeid=? AND dt >= $maxtime - $step AND dt < $maxtime", array($node['id'])); 
+		$download = ($data['download']?$data['download']:0);
+		$upload = ($data['upload']?$data['upload']:0);
+		if($download || $upload)
+		    $inserted += $LMS->DB->Execute("INSERT INTO stats (nodeid, dt, upload, download) VALUES (?, $maxtime, $upload, $download )", array($node['id']));
+		$maxtime -= $step;
+	    }
 	
-	$LMS->DB->CommitTrans();
-	echo ($deleted?$deleted:0)." - usuniêtych, ".($inserted?$inserted:0)." - wstawionych\n";
+	    $LMS->DB->CommitTrans();
+	    echo ($deleted?$deleted:0)." - usuniêtych, ".($inserted?$inserted:0)." - wstawionych\n";
+	}
     }
 }
 echo "Po kompaktowaniu w bazie pozostaje ".$LMS->DB->GetOne("SELECT COUNT(*) FROM stats")." rekordów.";
